+++ src/components/books/books/inputBook.js
@@ -1,7 +1,7 @@
 import { React, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { useDispatch } from 'react-redux';
-import { addBook } from '../../redux/books/books';
+import { addBook } from '../../../redux/books/books';
 
 const InputBook = () => {
   const [title, setTitle] = useState('');
