+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import App from './App';
-import store from './components/redux/configureStore';
+import store from './redux/configureStore';
 
 ReactDOM.render(
   <React.StrictMode>
